third_party/WebKit/LayoutTests/media/track/track-webvtt-tc030-interspersed-non-cue.html

<!DOCTYPE html>
<title>Tests that an empty line after an identifier line discards the current cue and restarts the cue loop.</title>
<script src="track-helpers.js"></script>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
check_cues_from_track("captions-webvtt/tc030-interspersed-non-cue.vtt", function(track) {
    var expected = [
        { text: "First" },
        { text: "Second" }
    ];

    assert_cues_match(track.cues, expected);
});
</script>
